NewsPortal/requirements.txt
django>=3.1
uvicorn>=0.15.0
asgiref>=3.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
requests==2.32.3
django-ninja
# mysqlclient>=2.1.0

